+++ GrokkingAlgorithms/EnumHelper.cs
@@ -1,0 +1,9 @@
+﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+
+namespace GrokkingAlgorithms
+{
+    public enum EnumWriteLine { True, False }
+    public enum EnumAlgorithm { First, Second, Third }
+    public enum EnumSort { Asc, Desc }
+}