+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.1.01] - 2019-09-30
+## [0.1.10] - 2019-09-30
 ### Added
 - .gitattributes
 - .gitignore
 - CHANGELOG.md
 - LICENSE
 - README.md
+- GrokkingAlgorithms solution & project
+- EnumHelper
+- ArrayHelper
+- Binary search algorithm.